refactor(src): separate gogoanime-specific logic
This separates gogoanime logic from the main entry point of the tool.
The purpose is to provide the groundwork for adding more providers in
the future.  Additionally, decrypt/extractor logic was moved out of the
main processing file for gogo.
