+++ src/providers/gogoanime/locales.ts
