+++ src/providers/index.ts
@@ -1,0 +1,5 @@
+import Gogoanime from "./gogoanime";
+
+export default {
+  Gogoanime: Gogoanime.initialize,
+};
